Backend/pyproject.toml
[project]
name = "carbon-reduction-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "geoalchemy2>=0.17.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "redis>=6.1.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn[standard]>=0.27.1",
    "bcrypt>=4.1.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'

[tool.flake8]
max-line-length = 88
extend-ignore = "E203"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv"
]
